bcachefs: set rebalance thread to SCHED_BATCH and nice 19
While the rebalance thread is usually not compute bound, it does cause
a considerable amount of I/O. Since "reducing" the nice level from 0
to 19, also implicitly reduces the thread's best-effort I/O scheduling
class level from 4 to 7. Hence, the rebalance thread's I/O will be
deprioritized over normal I/O.

Furthermore, we set the rebalance thread's scheduling class to BATCH,
which means it will potentially receive a higher scheduling
latency. Which is fine, given that it is not a latency sensitive task.

Signed-off-by: Florian Schmaus <flo@geekplace.eu>